feat: 네트워크 오류 분리
